Added ability to have group requirements (in degrees) with fractional hours.  Ex: 0.5 hours or 1.2 hours, etc.  Previously,   when a group was loaded, the hours_required field was being intval()'d.  Changed to floatval().
  Fixed bug where course descriptions were not being cached correctly due to a typo.

  Wrapped "hours" and "hrs" with t() function so those words can be translated to "credits" if need be in a hook_translate function.

+++ flightpath/classes/DegreePlan.php
@@ -654,8 +654,8 @@ class DegreePlan extends stdClass
         $group_g->assigned_to_semester_num = $semester_num;
         $group_g->title = "$title";
         $group_g->icon_filename = $icon_filename;
-        $group_g->hours_required = intval($cur["group_hours_required"]);
-        $group_g->min_hours_allowed = intval($cur["group_min_hours_allowed"]);
+        $group_g->hours_required = floatval($cur["group_hours_required"]);
+        $group_g->min_hours_allowed = floatval($cur["group_min_hours_allowed"]);
         $group_g->bool_placeholder = true;
         $obj_semester->list_groups->add($group_g);
